+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 * Player entities are now detected using a `IsPlayerEntity` boolean on their table
 * Fixed gamemode overriding of the `Musics` table
 * Fixed entity_burger_ai only targeting entity_burger (will now target any entity having .IsPlayerEntity = true)
+* Fixed weapon_graspain grapple force (is now based on entity mass)
 
 ## Beta 1.1